+++ bin/libs/movie.js
@@ -1,10 +1,12 @@
 /**
- * \libs\movie
+ * \libs\Movie
  *
  * @TODO: [✔] List
- * @TODO: [ ] Quality
+ * @TODO: [✔] Quality
+ * @TODO: [ ] Verify if is quality is set in arguments
  * @TODO: [ ] Subtitle
- * @TODO: [ ] Stream
+ * @TODO: [ ] Verify if language for subtitle is set
+ * @TODO: [✔] Stream
  */
 'use strict'
 
@@ -12,10 +14,18 @@ require('colors')
 
 const request = require('./helpers').request
     , list = require('./prompt').list
+    , subtitle = require('./subtitle')
+    , stream = require('./stream')
     , yts = require('../settings.json').yts
 
+/**
+ * Search movies.
+ * 
+ * @param  Object options
+ * @param  String q
+ */
 const getMovie = (options, q) => {
-  let coll = []
+  let movieList = []
   let url = yts.search + encodeURI(q) + '&sort_by=year&order_by=asc'
 
   request(url, response => {
@@ -26,16 +36,47 @@ const getMovie = (options, q) => {
     }
 
     response.map( e => {
-      coll.push({
+      movieList.push({
         name: e.title_long,
         value: e.id
       })
     })
 
-    list(coll, {
+    list(movieList, {
       name: 'id',
       message: 'Listing movies'
-    }, movie => console.log(movie.id))
+    }, movie => {
+      getQuality(movie.id)
+    })
+  })
+}
+
+/**
+ * List Availables qualities for movie.
+ * 
+ * @param  integer movieId
+ */
+const getQuality = movieId => {
+  let qualities = []
+  let url = yts.details + movieId
+
+  request(url, response => {
+    let movie = JSON.parse(response).data.movie
+    
+    movie.torrents.map( e => {
+      qualities.push({
+        name: e.quality,
+        value: e.url
+      })
+    })
+
+    list(qualities, {
+      name: 'url',
+      message: 'Available qualities'
+    }, q => {
+      subtitle({'imdbid': movie.imdb_code}, q.url)
+      // stream(movie.url)
+    })
   })
 }